requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchxrayvision>=0.0.39
PyQt6>=6.4.0
Pillow>=9.3.0
numpy>=1.21.0
monai>=1.2.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
